+++ app/src/main/java/com/mgc/sharesanalyse/entity/AvJsonBean.java
@@ -2,67 +2,67 @@ package com.mgc.sharesanalyse.entity;
 
 public class AvJsonBean {
 
-    private double AV;
-    private double AV100;
-    private double AV50;
-    private double AV30;
-    private double AV10;
-    private double AV5;
-    private double AV1;
+    private float AV;
+    private float AV100;
+    private float AV50;
+    private float AV30;
+    private float AV10;
+    private float AV5;
+    private float AV1;
 
-    public double getAV() {
+    public float getAV() {
         return AV;
     }
 
-    public void setAV(double AV) {
+    public void setAV(float AV) {
         this.AV = AV;
     }
 
-    public double getAV100() {
+    public float getAV100() {
         return AV100;
     }
 
-    public void setAV100(double AV100) {
+    public void setAV100(float AV100) {
         this.AV100 = AV100;
     }
 
-    public double getAV50() {
+    public float getAV50() {
         return AV50;
     }
 
-    public void setAV50(double AV50) {
+    public void setAV50(float AV50) {
         this.AV50 = AV50;
     }
 
-    public double getAV30() {
+    public float getAV30() {
         return AV30;
     }
 
-    public void setAV30(double AV30) {
+    public void setAV30(float AV30) {
         this.AV30 = AV30;
     }
 
-    public double getAV10() {
+    public float getAV10() {
         return AV10;
     }
 
-    public void setAV10(double AV10) {
+    public void setAV10(float AV10) {
         this.AV10 = AV10;
     }
 
-    public double getAV5() {
+    public float getAV5() {
         return AV5;
     }
 
-    public void setAV5(double AV5) {
+    public void setAV5(float AV5) {
         this.AV5 = AV5;
     }
 
-    public double getAV1() {
+    public float getAV1() {
         return AV1;
     }
 
-    public void setAV1(double AV1) {
+    public void setAV1(float AV1) {
         this.AV1 = AV1;
     }
 }
